sigma/1.1.0/requirements.txt
requests==2.32.4
pysigma==0.11.23
pysigma-backend-splunk==1.2.0
pysigma-backend-elasticsearch==1.1.3
pysigma-backend-kusto==0.4.0
pysigma-backend-qradar==0.4.0
pysigma-backend-logpoint==0.2.0
pysigma-backend-crowdstrike==0.3.0
pysigma-backend-carbonblack==0.2.0
pysigma-backend-insightidr==0.2.0
pysigma-backend-panther==0.2.0
pysigma-backend-opensearch==1.1.0
pysigma-backend-loki==1.0.0
pysigma-backend-sqlite==0.1.0
pyyaml==6.0.2
jsonschema==4.23.0
